+++ CHANGES.md
@@ -1,8 +1,13 @@
 # CHANGES
 
+## 1.5.0
+
+* IMPORT 1.0 changed to merge Seq Protocol and Data Type
+* Compitable with FastQs in CASAVA 1.8 format
+
 ## 1.4.2
 
-* added function to process `bz2` compressed FastQ format.  
+* added function to process `bz2` compressed FastQ format.
 
 ## 1.4.1